requirements.txt
Django==2.1.5
djangorestframework==3.10.2
pytz==2019.2
django-rest-swagger==2.1.2
django-reversion==2.0.9
django-permanent==1.1.6
djangorestframework-jwt==1.10.0
